refactor(tests): rename and update swap-specific-dex test
